webapp-main/requirements.txt
alembic==1.13.1
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
exceptiongroup==1.2.0
Flask==3.0.2
Flask-BasicAuth==0.2.0
Flask-Bcrypt==1.0.1
Flask-HTTPAuth==4.8.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-WTF==1.2.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-pubsub==2.21.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
packaging==23.2
passlib==1.7.4
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
PyMySQL==1.1.0
pytest==8.0.0
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
SQLAlchemy==2.0.27
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
WTForms==3.1.2
